Source code/backend/requirements.txt
Flask
psycopg2-binary
paho-mqtt
APScheduler
python-dotenv